Add nonce bytes to the blockID hash calculation

+++ nodeclient/http_api_client_test.go
@@ -404,7 +404,7 @@ func TestClient_CommitmentByID(t *testing.T) {
 
 	commitmentID := iotago.NewSlotIdentifier(slotIndex, tpkg.Rand32ByteArray())
 
-	commitment := iotago.NewCommitment(slotIndex, iotago.NewSlotIdentifier(slotIndex-1, tpkg.Rand32ByteArray()), tpkg.Rand32ByteArray(), int64(tpkg.RandUint64(math.MaxUint64)))
+	commitment := iotago.NewCommitment(slotIndex, iotago.NewSlotIdentifier(slotIndex-1, tpkg.Rand32ByteArray()), tpkg.Rand32ByteArray(), tpkg.RandUint64(math.MaxUint64))
 
 	data, err := v3API.Encode(commitment)
 	require.NoError(t, err)
@@ -451,7 +451,7 @@ func TestClient_CommitmentByIndex(t *testing.T) {
 
 	var slotIndex iotago.SlotIndex = 1337
 
-	commitment := iotago.NewCommitment(slotIndex, iotago.NewSlotIdentifier(slotIndex-1, tpkg.Rand32ByteArray()), tpkg.Rand32ByteArray(), int64(tpkg.RandUint64(math.MaxUint64)))
+	commitment := iotago.NewCommitment(slotIndex, iotago.NewSlotIdentifier(slotIndex-1, tpkg.Rand32ByteArray()), tpkg.Rand32ByteArray(), tpkg.RandUint64(math.MaxUint64))
 
 	data, err := v3API.Encode(commitment)
 	require.NoError(t, err)
